refactor: Sunset styled-components and styled-system
