+++ package.json
@@ -5,7 +5,6 @@
   "private": true,
   "dependencies": {
     "@phosphor-icons/react": "2.0.9",
-    "@styled-system/space": "5.1.2",
     "@supabase/supabase-js": "2.21.0",
     "@vercel/analytics": "1.0.0",
     "axios": "0.26.0",
@@ -20,7 +19,6 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-syntax-highlighter": "15.4.5",
-    "styled-components": "5.3.3",
     "swr": "2.1.5",
     "use-easing": "0.0.9"
   },
@@ -38,12 +36,9 @@
     "@testing-library/jest-dom": "5.16.5",
     "@testing-library/react": "14.0.0",
     "@types/node": "17.0.8",
-    "@types/nprogress": "0.2.0",
     "@types/react": "18.0.28",
     "@types/react-dom": "18.0.11",
     "@types/react-syntax-highlighter": "13.5.2",
-    "@types/styled-components": "5.1.20",
-    "@types/styled-system": "5.1.15",
     "@typescript-eslint/eslint-plugin": "5.10.0",
     "autoprefixer": "10.4.14",
     "csstype": "3.0.10",
@@ -62,8 +57,5 @@
     "postcss": "8.4.22",
     "tailwindcss": "3.3.1",
     "typescript": "4.9.5"
-  },
-  "msw": {
-    "workerDirectory": "public"
   }
 }
