--- src/@styled-system/space/index.d.ts
@@ -1,9 +1,0 @@
-import {
-  space as spaceFn,
-  SpaceProps as SpacePropsInterface
-} from "@types/styled-system";
-
-declare module "@styled-system/space" {
-  export const space: spaceFn;
-  export type SpaceProps = SpacePropsInterface;
-}
