--- src/components/Card.tsx
@@ -1,7 +1,0 @@
-import { BaseCard, Header, Section, CardProps } from "design-system/Card";
-
-export const Card = (props: CardProps) => <BaseCard {...props} />;
-
-Card.Header = Header;
-
-Card.Section = Section;
