+++ src/design-system/Background.tsx
@@ -15,29 +15,3 @@ export const Background = ({ readingMode, ...props }: BackgroundProps) => (
     )}
   />
 );
-// styled.div<BackgroundProps>`
-//   position: fixed;
-
-//   z-index: -1;
-
-//   top: 0;
-//   width: 100vw;
-//   height: 100vh;
-
-//   background-attachment: scroll;
-//   background-image: url("/waves.min.svg");
-//   background-size: cover;
-//   background-repeat: no-repeat;
-
-//   transform: ${(props) => props.readingMode && "translateY(20%)"};
-//   transition: transform 0.35s ease-in-out;
-//   will-change: auto;
-
-//   ${supportIsolate};
-
-//   & ~ * {
-//     z-index: 1;
-
-//     ${supportIsolate};
-//   }
-// `;
